apps/api/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-keycloak==3.7.0
cryptography==41.0.7

# Stripe
stripe==7.6.0

# OpenAI
openai==1.3.7

# External APIs
httpx==0.25.2
aiohttp==3.9.1

# Redis (rate limiting)
redis==5.0.1
aioredis==2.0.1

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
phonenumbers==8.13.26

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
